requirements.txt
celery==3.1.23
redis==2.10.5
tornado==4.4.2
tornado-celery==0.3.5
tornado-redis==2.4.18
